Updated WiFi configuration to remove security vulnerability on wireless boards. Updated pinout list to only include valid input/output pins. Added comments.
Updates to the original repository from FutureTechGuy to shut down access point mode on the WiFi circuit, which was allowing anyone to connect to the network. Also tested the GPIO pins with hardware documentation to remove invalid pins for triggers and lights, moving the "presspin" LED to a better GPIO pin and allowing for the maximum number of function buttons to be used. Comments included to guide customization and necessary edits.

Co-Authored-By: KinzuaKid <64295066+kinzuakid@users.noreply.github.com>
Co-Authored-By: Wendell Seaton <futuretechguy@gmail.com>